+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `AccountSettings.documentsMerging`
 - Add `MyEnvelope.subjectName`
 - Add `MyEnvelope.validTo`
+- Add `AccountIdentifyScenariosEndpoint` and `IdentifyScenario` resource
+- Add `AccountIdentifyScenarioVersionsEndpoint` and `IdentifyScenarioVersion` resource
 
 ### Changed
 - Change `Identification.envelope` from string to EnvelopeInfo